+++ CHANGELOG.md
@@ -31,7 +31,8 @@ For changes to the BPDM Helm charts please consult the [changelog](charts/bpdm/C
 - BPDM Cleaning Service Dummy: Added a null check for name parts to ensure proper whitespace handling when constructing the legal name from them.
 - BPDM Gate: Enabled Tax Jurisdiction code to save it to the Output.
 - BPDM Cleaning Service Dummy: Removed assignment of uncategorized states while performing cleaning legal entity process.
-- BPDM Gate: Fixed construction logic for states and identifiers by enabling business partner type 
+- BPDM Gate: Fixed construction logic for states and identifiers by enabling business partner type
+- BPDM Gate: Fixed logic for identifiers to retrieve only generic type on output business partner
 
 ## [6.1.0] - [2024-07-15]